requirement.txt
asttokens==2.4.1
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
colorama==0.4.6
einops==0.7.0
executing==2.0.1
filelock==3.9.0
h5py==3.11.0
icecream==2.1.3
idna==3.4
imageio==2.34.0
Jinja2==3.1.2
joblib==1.4.0
lit==15.0.7
MarkupSafe==2.1.3
MedPy==0.5.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.8.0.76
pillow==10.2.0
Pygments==2.17.2
requests==2.28.1
safetensors==0.3.3
scikit-learn==1.4.2
scipy==1.13.0
SimpleITK==2.3.1
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
torch==2.0.1+cu118
torchsummary==1.5.1
torchvision==0.15.2+cu118
tqdm==4.66.2
triton==2.0.0
typing_extensions==4.8.0
urllib3==1.26.13
